# docs/modules/ROOT/attachments/openapi.yaml
openapi: 3.1.0
info:
  title: JSupervisor REST API
  description: |
    JSupervisor is a process management and monitoring service that provides
    comprehensive control over application processes. This API enables you to:

    - Start, stop, and restart processes individually or in bulk
    - Monitor process status and health
    - Retrieve process logs and statistics
    - Configure process behavior and auto-start settings

    The API follows RESTful principles and returns JSON responses.
  contact:
    name: Alex Mondshain
    url: https://www.alexmond.org/jsupervisor/current/
    email: alex.mondshain@gmail.com
  license:
    name: Apache License 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.txt
  version: 0.0.2
servers:
  - url: http://localhost:8086
    description: Local Development Server
...
paths:
...
  /api/v1/processes/stop/{name}:
    post:
      tags:
        - Process Management
      summary: Stop process
      description: Stop a specific process
      operationId: stopProcess
      parameters:
        - name: name
          in: path
          description: Process name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Process stop initiated successfully
        "404":
          description: Process not found
        "500":
          description: Failed to stop process
  /api/v1/processes/start/{name}:
    post:
      tags:
        - Process Management
      summary: Start process
      description: Start a specific process
      operationId: startProcess
      parameters:
        - name: name
          in: path
          description: Process name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Process start initiated successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseMessage"
        "400":
          description: Invalid process configuration
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseMessage"
        "500":
          description: Failed to start process
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseMessage"
  /api/v1/processes/restart/{name}:
    post:
      tags:
        - Process Management
      summary: Restart process
      description: Restart a specific process
      operationId: restartProcess
      parameters:
        - name: name
          in: path
          description: Process name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Process restart initiated successfully
        "404":
          description: Process not found
        "500":
          description: Failed to restart process
...
  /api/v1/processes/status/{name}:
    get:
      tags:
        - Process Management
      summary: Get process status
      description: Get the current status of a specific process
      operationId: getProcessStatus
      parameters:
        - name: name
          in: path
          description: Process name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successfully retrieved process status
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProcessStatusRest"
        "404":
          description: Process not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProcessStatusRest"
        "500":
          description: Failed to retrieve process status
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProcessStatusRest"
...
components:
  schemas:
...
    ResponseMessage:
      type: object
      properties:
        message:
          type: string
        error:
          type: string
...
    ProcessStatusRest:
      type: object
      description: Represents process status information
      properties:
        name:
          type: string
          description: Name of the process
          example: myapp
        status:
          type: string
          description: "Current status of the process (e.g., running, stopped, failed)"
          enum:
            - not_started
            - running
            - finished
            - unknown
            - failed
            - failed_to_start
            - stopped
            - stopping
            - aborted
            - starting
            - healthy
            - unhealthy
        pid:
          type: integer
          format: int64
          description: Process ID assigned by the operating system
          example: 1234
        startTime:
          type: string
          format: date-time
          description: Timestamp when the process was started
        endTime:
          type: string
          format: date-time
          description: Timestamp when the process ended
        exitCode:
          type: integer
          format: int32
          description: Process exit code
          example: 0
        processRuntime:
          type: string
          format: duration
          description: Process runtime duration
        processUptime:
          type: string
          description: Formatted process uptime
          example: 2d 5h 30m
        stdoutLogfile:
          type: string
          description: "File path where the standard output (stdout) of the process\
            \ will be logged. If not specified, stdout will be inherited from the\
            \ parent process."
        stderrLogfile:
          type: string
          description: "File path where the standard error (stderr) of the process\
            \ will be logged. If not specified, stderr will be inherited from the\
            \ parent process."
        failedErrorLog:
          type: string
          description: Error log content when process failed to start
        alive:
          type: boolean
...